+++ TestFabricMod/src/main/java/io/github/natanfudge/nectest/NecTestCrash.java
@@ -1,4 +1,4 @@
-package io.github.natanfudge;
+package io.github.natanfudge.nectest;
 
 public class NecTestCrash extends RuntimeException{
     public NecTestCrash(String message){
