+++ TestFabricMod/src/main/resources/nec_testmod.mixins.json
@@ -1,13 +1,13 @@
 {
   "required": true,
   "minVersion": "0.8",
-  "package": "io.github.natanfudge.mixin",
+  "package": "io.github.natanfudge.nectest.mixin",
   "compatibilityLevel": "JAVA_8",
   "mixins": [
-    "ExampleMixin",
     "MixinMinecraftServer"
   ],
   "client": [
+    "ExampleMixin"
   ],
   "injectors": {
     "defaultRequire": 1
