+++ TestForgeMod/src/main/resources/nec_testmod.mixins.json
@@ -4,10 +4,10 @@
   "package": "io.github.natanfudge.nectest.mixin",
   "compatibilityLevel": "JAVA_16",
   "mixins": [
-    "ExampleMixin",
     "MixinMinecraftServer"
   ],
   "client": [
+    "ExampleMixin"
   ],
   "injectors": {
     "defaultRequire": 1
