+++ gradle.properties
@@ -2,20 +2,21 @@ org.gradle.jvmargs=-Xmx2048M
 
 
 archives_base_name=notenoughcrashes
-mod_version=4.1.0
+mod_version=4.1.1
 maven_group=fudge.notenoughcrashes
 
-architectury_version=1.8.131
+architectury_version=2.5.28
 
-forge_version=36.0.1
 
-minecraft_version=1.16.5
-yarn_mappings=1.16.5+build.3
+forge_version=37.0.69
+
+minecraft_version=1.17.1
+yarn_mappings=1.17.1+build.61
 loader_version=0.12.5
 
 #Fabric api
-fabric_version=0.29.4+1.16
+fabric_version=0.40.1+1.17
 
 cursegradle_version = 1.4.0
-curseforge_mc_version=1.16.5
+curseforge_mc_version=1.17.1
 mod_name=Not Enough Crashes
